requirements.txt
argparse
easydict
h5py
matplotlib
numpy
open3d==0.9.0.0
opencv-python
scipy
tensorboardX==1.2
transforms3d
tqdm
einops
munch
termcolor
pymesh
pyyaml
tensorboard